+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased] - 2022-05-05
+- Fixed: missing AdvancedUserInterface in symfony 5
+
 ## [2.2.1] - 2022-05-04
 - Fixed: uitls bundle dependency